AST

Description

AST encodes the different datatypes used throughout the application.

Documentation

data Command

Encoding of theater keywords for intruction of operators.

Constructors

Go
Instructs operator should take instruction immediately.

Stby
Instructs operator to get ready and confirm ready.

Warn
A polite/stategic reminder for some instruction.

data Character

An encoding of a physical Character that could appear in a script, cue, etc.

Constructors

Character

Fields

charName :: String

data Action

An encoding of a physical action that can be taken.

Constructors

Enter

Exit

data Department

A Department corresponds to a physical department in the theater. (e.g. Lights, Sound, Backstage, etc.)

Constructors

Department

Fields

deptName :: String

data Cue

A Cue is the smallest unit of action in our representation. This representation has no notion of time/placement by design and must be coupled else where (e.g CueGroup which has a Marker specifying time.)

Constructors

Cue

Fields

department :: Department

number :: Int

command :: Command

comment :: Maybe String

data CueGroup

A binding of a Marker and list of cues to occur at that Marker.

Constructors

CueGroup

Fields

cgMarker :: Marker

cgCues :: [Cue]

data Marker

A Marker is a (possibly partial) reference to a part of a play.

Constructors

Visual

Fields

mChar :: Character

mAction :: Action

mIdx :: Maybe Int

Line

Fields

mChar :: Character

mLine :: String

mIdx :: Maybe Int

data CueScene

Internal representation of a scene and any groups of cues that appear throughout it.

Constructors

CueScene

Fields

csIndex :: Int

csGroups :: [CueGroup]

data CueAct

Internal represenation of an act and its scenes.

Constructors

CueAct

Fields

cueActIdx :: Int

cueActScenes :: [CueScene]

data CueSheet

Internal representation of a cue sheet: characters, departments, and acts.

Constructors

CueSheet

Fields

characters :: [Character]

departments :: [Department]

acts :: [CueAct]

data PromptScript

Analagous to CueSheet for use in `Prompt*` using a Map internally for easier traversals, lookups, and guarentees uniqueness as well as Sets.

Constructors

PromptScript

Fields

pTitle :: String

pCharacters :: Set Character

pDepartments :: Set Department

pActs :: Map Int PromptAct

data PromptAct

Analagous to CueAct for use in `Prompt*` using a Map internally for easier traversals, lookups, and guarentees uniqueness.

Constructors

PromptAct

Fields

pActScenes :: Map Int PromptScene

data PromptScene

Analagous to CueScene for use in `Prompt*` using a Map internally for easier traversals, lookups, and guarentees of uniqueness.

Constructors

PromptScene

Fields

pMarkers :: Map Int PromptMarker

data PromptMarker

Analogous to CueMarker for use in `Prompt*`.

Constructors

PromptMarker

Fields

pMarker :: Marker

pCues :: [Cue]