Interp

Description

Interpreter for the DSL to generate PromptScript's.

Documentation

transpile :: CueSheet -> Either Error PromptScript

Compile a CueSheet into a PromptScript safely.

insertAct :: PromptScript -> CueAct -> Either Error PromptScript

Inserts a CueAct into the appropriately place in a PromptScript.

insertScene

Arguments

:: Int
Act index of target.

-> PromptScript
Full PromptScript to be updated.

-> CueScene
CueScene to be inserted.

-> Either Error PromptScript
Updated PromptScript or Error.

Inserts a CueScene into a PromptScript.

checkUnkownCharacters

Arguments

:: PromptScript
Script

-> PromptScript
Cue Sheet

-> Either [Error] ()
Result

Check unkown characters.

eval

Arguments

:: PromptScript
The prompt script, possibly already marked up.

-> PromptScript
The compiled cuesheet.

-> Either [Error] PromptScript
A list of Errors, or the updated prompt script.

Merge a prompt script with a cuesheet.

processAct

Arguments

:: PromptScript
The prompt script to be updated.

-> [Error]
Accumulating errors.

-> Int
Act index.

-> PromptAct
Act from cue sheet to be merged in.

-> ([Error], PromptAct)
Any errors and updated act.

Merge PromptActs scenes into PromptScript.

processScene

Arguments

:: PromptScript
The cuehseet

-> Int
Act index

-> [Error]
Accumulating errors

-> Int
Scene index

-> PromptScene
Original scene

-> ([Error], PromptScene)
Results

Merge PromptSceness cue groups into PromptScript.

resolveMarker

Arguments

:: ([Error], Map Int PromptMarker)
Any errors and the "script".

-> PromptMarker
PromptMarker to be placed.

-> (([Error], Map Int PromptMarker), Map Int PromptMarker)
First is updated, IGNORE snd.

Attempt to place a PromptMarker into a PromptScript.

insertCueGroups

Arguments

:: Int
Act index.

-> Int
Scene index.

-> PromptScript
Prompt script to be updated.

-> CueGroup
CueGroup to place

-> Either Error PromptScript
Updated PromptScript or Error.

Attempt to insert a CueGroup into a script.

insertPromptMarker

Arguments

:: Int
Act index.

-> Int
Scene index.

-> PromptMarker
Marker to be inserted.

-> PromptScript
Prompt script to be updated.

-> Either e PromptScript
Updated prompt script.

Insert a PromptMarker into the script UNSAFELY.

genKey :: (Enum k, Bounded k) => Map k v -> k

Generator a fresh key to be used in a Map based on the highest index.

safeListToSet :: (Ord a, Traversable t) => (a -> e) -> t a -> Either e (Set a)

Safely convert a list of elements to a set, complaining if there are any duplicates.

ensureStrictGtInsert :: Ord k => (k -> e) -> k -> Map k v -> Either e (Map k v)

Check to see if the item attempting to be inserted, is strictly greater than the highest element in the map.

fuzzyMatch

Arguments

:: PromptMarker
Search Criteria

-> Int
Marker Index

-> PromptMarker
Possible match

-> Bool
True if fuzzy match, False otherwise.

Fuzzy check if two PromptMarkers match.